+++ docs/changes.rst
@@ -20,6 +20,9 @@ Released: not yet
   b) incompatibility between click 8.0 and clicl-repl.
   (see issues #816 and #817)
 
+* Limit mock package to lt 4.0.3 to avoid issue issue that causes test failure.
+  (see #822)
+
 **Enhancements:**
 
 **Known issues:**