+++ docs/changes.rst
@@ -30,6 +30,9 @@ Released: not yet
 
 **Enhancements:**
 
+* Migrated from Travis and Appveyor to GitHub Actions. This required several
+  changes in package dependencies for development.
+
 **Known issues:**
 
 * See `list of open issues`_.