+++ docs/changes.rst
@@ -5,6 +5,28 @@ Change log
 ==========
 
 
+pywbemtools 0.8.2.dev1
+^^^^^^^^^^^^^^^^^^^^^^
+
+Released: not yet
+
+**Incompatible changes:**
+
+**Deprecations:**
+
+**Bug fixes:**
+
+**Enhancements:**
+
+**Cleanup:**
+
+**Known issues:**
+
+* See `list of open issues`_.
+
+.. _`list of open issues`: https://github.com/pywbem/pywbemtools/issues
+
+
 pywbemtools 0.8.1
 -----------------